* moved functions out into a new utils module to prepare to drop them into DataVisualizer
